requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
